Move setup into own command, upgrade clap
