+++ Cargo.toml
@@ -11,7 +11,7 @@ edition = "2018"
 
 [dependencies]
 cargo_metadata = "0.11.0"
-clap = { version = "3.1.12", features = ["derive"] }
+clap = { version = "4.1.8", features = ["derive"] }
 goblin = "0.2.3"
 serde = "1.0"
 serde_derive = "1.0"
